13/02/2024 21:51

+++ AprobarInesMR/GraphQL/src/main/java/com/example/graphql/domain/modelo/Rol.java
@@ -1,3 +1,3 @@
 package com.example.graphql.domain.modelo;
 
-public record Rol (int id, String rol) {}
+public record Rol (Long id, String rol) {}
