+++ AprobarInesMR/GraphQL/src/main/java/com/example/graphql/domain/modelo/User.java
@@ -2,7 +2,7 @@ package com.example.graphql.domain.modelo;
 
 import java.util.List;
 
-public record User(int id,
+public record User(Long id,
                    String username,
                    String password,
                    List<Rol> roles) {
